+++ FinalProject/src/main/resources/templates/chat/chatroom.html
@@ -50,7 +50,6 @@
                 <div class="input-group clearfix">
                     <!-- <img src="/image/chat/none.png" alt="이미지 없음"/> -->
                     
-                    
                     <input type="text" id="message" placeholder="" autocomplete="off"
                            class="form-control messageInput"/>
                     <button type="submit" class="primary">Send</button>
